Add files via upload
This is still very similar to the model in the file 'model.ipynb'. There are some additions, such as the operational rule and the definition (random assignment) of more parameters and variables. 
I have re-written the codes mainly as an exercise for me to go through the basic p2p model and understand how it works by unpacking it down to each single step.